requirements.txt
# Django
Django>=5.0,<6.0
django-modeltranslation>=0.19.0
django-unfold>=0.36.0
django-environ>=0.11.0
gunicorn>=22.0.0
whitenoise>=6.0.0

# API
django-ninja>=1.3.0

# Database
psycopg2-binary>=2.9.0

# Images
Pillow>=10.0.0
